Fix issue with nested ConstraintLayout in studio - on android, CL doesn't do nesting (i.e. nested CLs are for now just solved one after the other) - but in Studio, CLs are solved as part of the same tree, which causes some issues
The fix makes sure to snapshot, reset, and reapply the snapshot on nested CLs. This would be a no-op on android (mParent would be null), but replicates the solving in isolation in Studio. The second part of the fix is in WidgetContainer and don't update Draw positions for more than one layer of CLs.