Fix issue with nested ConstraintLayout in studio - on android, CL doesn't do nesting (i.e. nested CLs are for now just solved one after the other) - but in Studio, CLs are solved as part of the same tree, which causes some issues
The fix makes sure to snapshot, reset, and reapply the snapshot on nested CLs. This would be a no-op on android (mParent would be null), but replicates the solving in isolation in Studio. The second part of the fix is in WidgetContainer and don't update Draw positions for more than one layer of CLs.
(cherry picked from commit 3b1045cb0f02c456e18912a1946ace68f80a624d)

Also Fix padding support in solver

Fixes: https://code.google.com/p/android/issues/detail?id=221690

Change-Id: If77c8687b3ff1e6bf441a653d1150f9b5e0e60cc