Add packed chain behavior.
A packed chain only work if elements of a chain are in fixed size (WRAP_CONTENT or set size).
If a chain is defined as packed, its position will be controlled by the corresponding dimension
bias of the first element of the chain. By default, elements of a packed chain will thus be centered.